casted mysqli_insert_id to int in getLastId()
git-svn-id: https://quaaxtm.svn.sourceforge.net/svnroot/quaaxtm@236 11aa7465-d437-4035-867d-07b0d7ee7c12

+++ branches/apichange/src/utils/MysqlResult.class.php
@@ -84,7 +84,7 @@ class MysqlResult {
    * @return int
    */
   public function getLastId() {
-    return mysqli_insert_id($this->connection);
+    return (int) mysqli_insert_id($this->connection);
   }
 }
 ?>
